app/requirements.txt
mysql-connector-python==8.0.33
langchain==0.0.219
pandas==2.0.3 
